Made ancestors & relcache more portable.
This makes it possible to switch between 32-bit and 64-bit tundra
binaries. Also, rather than exiting when an invalid ancestor file is
detected, tundra will just rebuild everything which much better.

Closes GH-105 (Support mixing 32-bit and 64-bit ancestor files)

+++ src/engine.c
@@ -370,7 +370,7 @@ static int make_engine(lua_State *L)
 
 	self->file_hash_size = 92413;
 	self->relhash_size = 92413;
-	self->start_time = time(NULL);
+	self->start_time = (uint64_t) time(NULL);
 	self->settings.thread_count = td_get_processor_count();
 
 	/* apply optional overrides */
